+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # v3.0.3
 * Added: Added a hasAuthorization method to the Query class to allow searches for any element that uses an authorization string or strings.
 * Added: Query extended data on an element
+* Added: EmptyResultsGraphQuery and EmptyResultsQueryResultsIterable
 * Fixed: GeoPoint.distanceBetween and GeoPoint.calculateCenter calculations
 * Fixed: Elasticsearch5 field limit of 1000
+* Removed: CompositeGraphQuery
 
 # v3.0.2
 * Fixed: Elasticsearch5 throwing unsupported operation exception when adding/updating a property